rename exception to exceptions
