+++ src/bmkg/exceptions.py
