+++ src/bmkg/parsers/parse_datetime_element.py
@@ -5,7 +5,7 @@ from datetime import datetime
 # Element is used only for typing not parsing
 from xml.etree.ElementTree import Element  # nosec B405
 
-from ..exception import WeatherForecastParseError
+from ..exceptions import WeatherForecastParseError
 
 __all__ = ["parse_datetime_element"]
