+++ src/bmkg/parsers/parse_temperature_element.py
@@ -4,7 +4,7 @@ from collections.abc import Iterator
 # Element is used only for typing not parsing
 from xml.etree.ElementTree import Element  # nosec B405
 
-from ..exception import WeatherForecastParseError
+from ..exceptions import WeatherForecastParseError
 from ..schemas import Temperature
 
 __all__ = ["parse_temperature_element"]
